+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `<cac:OriginatorDocumentReference>` to `<Invoice>`
 - Add FRCTC Electronic Address (0225) to EASCode - Thanks [@UbiManu](https://github.com/UbiManu)
 - Add `<cac:AddressLine>` support to `<cac:Address>` - Thanks [@fMads](https://github.com/fMads)
+- Add `<cac:OriginCountry>` support to `<cac:Item>` - Thanks [@fMads](https://github.com/fMads)
 
 ### Bug fixes